fixed small bug

+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   github:   https://github.com/priezu-m                                    */
 /*   Licence:  GPLv3                                                          */
 /*   Created:  2023/10/04 01:13:48                                            */
-/*   Updated:  2023/10/07 20:34:02                                            */
+/*   Updated:  2023/10/13 18:07:26                                            */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define LIBFT_H
 
 # include <stddef.h>
+# include <stdint.h>
 
 ;
 # pragma clang diagnostic push
@@ -26,7 +27,7 @@ size_t	ft_strlen(const char *str);
 int		ft_strcmp(const char *s1, const char *s2);
 int		ft_isspace(char c);
 void	ft_putstr_fileno(int fileno, const char *str);
-void	ft_putnbr_fileno(int fileno, int nbr);
+void	ft_putnbr_fileno(int fileno, uintmax_t nbr);
 
 # pragma clang diagnostic pop
